move receiver into an internal package, in preparation for profiles
