+++ receiver/internal/traces.go
@@ -1,0 +1,15 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package internal // import "go.opentelemetry.io/collector/receiver/internal"
+
+import "go.opentelemetry.io/collector/component"
+
+// Traces receiver receives traces.
+// Its purpose is to translate data from any format to the collector's internal trace format.
+// TracesReceiver feeds a consumer.Traces with data.
+//
+// For example, it could be Zipkin data source which translates Zipkin spans into ptrace.Traces.
+type Traces interface {
+	component.Component
+}
